% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_collect.R
\name{cluster_collect}
\alias{cluster_collect}
\title{Collect pairs from cluster nodes}
\usage{
cluster_collect(pairs, clear = FALSE)
}
\arguments{
\item{pairs}{an object or type \code{cluster_pairs} as created for example by
\code{\link{cluster_pair}}.}

\item{clear}{remove the pairs from the cluster nodes}
}
\value{
Returns an object of type \code{pairs} which is a \code{data.table}. This
object can be used as a regular (non-cluster) set of pairs
}
\description{
Collect pairs from cluster nodes
}
\examples{
library(parallel)
data("linkexample1", "linkexample2")
cl <- makeCluster(2)
pairs <- cluster_pair(cl, linkexample1, linkexample2)
local_pairs <- cluster_collect(pairs, clear = TRUE)
stopCluster(cl)

}
