% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_threshold.R
\name{select_threshold}
\alias{select_threshold}
\title{Select matching pairs with a score above a threshold}
\usage{
select_threshold(pairs, variable, score, threshold, inplace = FALSE, ...)
}
\arguments{
\item{pairs}{a \code{pairs} object, such as generated by 
\code{\link{pair_blocking}}}

\item{variable}{the name of the new variable to create in pairs. This will be a
logical variable with a value of \code{TRUE} for the selected pairs.}

\item{score}{name of the score/weight variable of the pairs. When not given
and \code{attr(pairs, "score")} is defined, that is used.}

\item{threshold}{the threshold to apply. Pairs with a score above the 
threshold are selected.}

\item{inplace}{logical indicating whether \code{pairs} should be modified in place. When
pairs is large this can be more efficient.}

\item{...}{ignored}
}
\value{
Returns the \code{pairs} with the variable given by \code{variable} added. This
is a logical variable indicating which pairs are selected a matches.
}
\description{
Select matching pairs with a score above a threshold
}
\examples{
data("linkexample1", "linkexample2")
pairs <- pair_blocking(linkexample1, linkexample2, "postcode")
pairs <- compare_pairs(pairs, c("lastname", "firstname", "address", "sex"))
model <- problink_em(~ lastname + firstname + address + sex, data = pairs)
pairs <- predict(model, pairs, type = "mpost", add = TRUE, binary = TRUE)

# Select pairs with a mpost > 0.5
pairs <- select_threshold(pairs, "selected", "mpost", 0.5, inplace = TRUE)

}
