% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{val_or}
\alias{val_or}
\title{Return default value if value is missing or NULL}
\usage{
val_or(val, or)
}
\arguments{
\item{val}{value to check}

\item{or}{value to return if \code{val} is missing or \code{NULL}.}
}
\value{
Returns \code{or} is val is \code{NULL}. Otherwise it returns
\code{ifelse(is.na(val), or, val)}.
}
\description{
Return default value if value is missing or NULL
}
