% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_from_x.R
\name{add_from_x}
\alias{add_from_x}
\alias{add_from_y}
\title{Add a variable from one of the data sets to pairs}
\usage{
add_from_x(pairs, variable, new_variable = variable, ...)

add_from_y(pairs, variable, new_variable = variable, ...)
}
\arguments{
\item{pairs}{\code{\link{data.table}} with pairs. Should contain the columns 
\code{.x} and \code{.y}.}

\item{variable}{name of the variable that should be added}

\item{new_variable}{optional variable name of the new variable in
\code{pairs}. When omitted \code{variable} is used.}

\item{...}{other parameters are passed on to \code{compare_vars}. Especially
\code{inplace}, \code{x} and \code{y} might be of interest.}
}
\value{
Returns the pairs with the column added. When \code{inplace = TRUE}
\code{pairs} is returned invisibly and the original \code{pairs} is 
modified.
}
\description{
Add a variable from one of the data sets to pairs
}
